// src/types.ts
export type UserRole = 'FARMER' | 'CITIZEN' | 'OFFICER' | 'ADMIN';

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  location: string;
  avatar?: string;
  farmSizeAcres?: number;
}